+++ .gitignore
@@ -8,3 +8,5 @@ __pycache__/
 # Data / misc
 .ipynb_checkpoints/
 *.json
+
+tweets.txt
