--- test_stream_listener.py
@@ -1,9 +1,0 @@
-from stream_listener import SListener
-
-
-def test():
-    listener = SListener()
-
-
-if __name__ == '__main__':
-    test()
